refactor(triple-web): modal을 internal-components로 이동
