+++ packages/triple-web/src/components/index.ts
@@ -1,4 +1,2 @@
 export * from './event-metadata-provider'
 export * from './event-tracking-provider'
-export * from './login-cta-modal'
-export * from './transition-modal'
